requirements.txt
pypdf
langchain
chromadb # Vector storage
pytest
openai
langchain-community
langchain-openai
markdown
streamlit
python-dotenv
langchain.tools
langchain-chroma
pandas
rank-bm25